combine KeyState and IKeyInfo move IKeyInfo to common
