--- Application/src/main/java/io/github/antijava/marjio/input/IKeyInfo.java
@@ -1,13 +1,0 @@
-package io.github.antijava.marjio.input;
-
-import io.github.antijava.marjio.common.input.Key;
-
-/**
- * Created by firejox on 2015/12/25.
- */
-
-
-public interface IKeyInfo {
-    Key getKey();
-    KeyState getKeyState();
-}
