--- Application/src/main/java/io/github/antijava/marjio/input/KeyState.java
@@ -1,10 +1,0 @@
-package io.github.antijava.marjio.input;
-
-/**
- * Created by firejox on 2015/12/25.
- */
-public enum KeyState {
-    KEY_PRESSED,
-    KEY_RELEASED,
-    KEY_UNKNOWN,
-}
