+++ Common/src/main/java/io/github/antijava/marjio/common/input/Event.java
@@ -3,9 +3,22 @@ package io.github.antijava.marjio.common.input;
 /**
  * Created by firejox on 2015/12/25.
  */
-public abstract class Event {
-    public abstract Type getType();
-    public abstract Object getData();
+public class Event {
+    Object obj;
+    Type type;
+
+    Event (Object obj, Type type) {
+        this.obj = obj;
+        this.type = type;
+    }
+
+    public Type getType() {
+        return type;
+    }
+
+    public Object getData() {
+        return obj;
+    }
 
     public enum Type {
         Keyboard,
